工作流库0216/T38-年报处理/requirements.txt
# T38-年报处理 依赖包

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0

# PDF处理
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# 网络请求
requests>=2.31.0
urllib3>=2.0.0

# 大模型
openai>=1.0.0

# 环境变量管理
python-dotenv>=1.0.0

# Excel导出
openpyxl>=3.1.0

# 可选：百度OCR
# baidu-aip>=4.16.0
